refactor: move tests to e2e subfolder
